+++ example/macos/Runner/MainFlutterWindow.swift
@@ -1,18 +1,18 @@
 import Cocoa
 import FlutterMacOS
-import flutter_acrylic
+import macos_window_utils
 
 class MainFlutterWindow: NSWindow {
   override func awakeFromNib() {
     let windowFrame = self.frame
-    let blurryContainerViewController = BlurryContainerViewController() // new
-    self.contentViewController = blurryContainerViewController // new
+    let macOSWindowUtilsViewController = MacOSWindowUtilsViewController()
+    self.contentViewController = macOSWindowUtilsViewController
     self.setFrame(windowFrame, display: true)
-    
-    /* Initialize the flutter_acrylic plugin */
-    MainFlutterWindowManipulator.start(mainFlutterWindow: self) // new
-    
-    RegisterGeneratedPlugins(registry: blurryContainerViewController.flutterViewController) // new
+
+    /* Initialize the macos_window_utils plugin */
+    MainFlutterWindowManipulator.start(mainFlutterWindow: self)
+
+    RegisterGeneratedPlugins(registry: macOSWindowUtilsViewController.flutterViewController)
 
     super.awakeFromNib()
   }