+++ lib/window.dart
@@ -3,10 +3,14 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_acrylic/macos/converters/blur_view_state_to_visual_effect_view_state_converter.dart';
+import 'package:flutter_acrylic/macos/converters/mac_toolbar_style_to_window_toolbar_style_converter.dart';
 import 'package:flutter_acrylic/macos/macos_blur_view_state.dart';
 import 'package:flutter_acrylic/macos/macos_toolbar_style.dart';
 import 'package:flutter_acrylic/macos/visual_effect_view_properties.dart';
+import 'package:flutter_acrylic/macos/converters/window_effect_to_material_converter.dart';
 import 'package:flutter_acrylic/window_effect.dart';
+import 'package:macos_window_utils/window_manipulator.dart';
 
 /// Platform channel name.
 const _kChannelName = "com.alexmercerind/flutter_acrylic";
@@ -29,152 +33,6 @@ const _kEnterFullscreen = "EnterFullscreen";
 /// Exits fullscreen.
 const _kExitFullscreen = "ExitFullscreen";
 
-/// Gets the height of the titlebar.
-const _kGetTitlebarHeight = "GetTitlebarHeight";
-
-/// Overrides the brightness setting of the window (macOS only).
-const _kOverrideMacOSBrightness = "OverrideMacOSBrightness";
-
-/// (macOS only).
-const _kSetDocumentEdited = "SetDocumentEdited";
-
-/// (macOS only).
-const _kSetDocumentUnedited = "SetDocumentUnedited";
-
-/// (macOS only).
-const _kSetRepresentedFile = "SetRepresentedFile";
-
-/// (macOS only).
-const _kSetRepresentedURL = "SetRepresentedURL";
-
-/// (macOS only).
-const _kHideTitle = "HideTitle";
-
-/// (macOS only).
-const _kShowTitle = "ShowTitle";
-
-/// (macOS only).
-const _kMakeTitlebarTransparent = "MakeTitlebarTransparent";
-
-/// (macOS only).
-const _kMakeTitlebarOpaque = "MakeTitlebarOpaque";
-
-/// (macOS only).
-const _kEnableFullSizeContentView = "EnableFullSizeContentView";
-
-/// (macOS only).
-const _kDisableFullSizeContentView = "DisableFullSizeContentView";
-
-/// (macOS only).
-const _kZoomWindow = "ZoomWindow";
-
-/// (macOS only).
-const _kUnzoomWindow = "UnzoomWindow";
-
-/// (macOS only).
-const _kIsWindowZoomed = "IsWindowZoomed";
-
-/// (macOS only).
-const _kIsWindowFullscreened = "IsWindowFullscreened";
-
-/// (macOS only).
-const _kHideZoomButton = "HideZoomButton";
-
-/// (macOS only).
-const _kShowZoomButton = "ShowZoomButton";
-
-/// (macOS only).
-const _kHideMiniaturizeButton = "HideMiniaturizeButton";
-
-/// (macOS only).
-const _kShowMiniaturizeButton = "ShowMiniaturizeButton";
-
-/// (macOS only).
-const _kHideCloseButton = "HideCloseButton";
-
-/// (macOS only).
-const _kShowCloseButton = "ShowCloseButton";
-
-/// (macOS only).
-const _kEnableZoomButton = "EnableZoomButton";
-
-/// (macOS only).
-const _kDisableZoomButton = "DisableZoomButton";
-
-/// (macOS only).
-const _kEnableMiniaturizeButton = "EnableMiniaturizeButton";
-
-/// (macOS only).
-const _kDisableMiniaturizeButton = "DisableMiniaturizeButton";
-
-/// (macOS only).
-const _kEnableCloseButton = "EnableCloseButton";
-
-/// (macOS only).
-const _kDisableCloseButton = "DisableCloseButton";
-
-/// (macOS only).
-const _kIsWindowInLiveResize = "IsWindowInLiveResize";
-
-/// (macOS only).
-const _kSetWindowAlphaValue = "SetWindowAlphaValue";
-
-/// (macOS only).
-const _kIsWindowVisible = "IsWindowVisible";
-
-/// (macOS only).
-const _kSetWindowBackgroundColorToDefaultColor =
-    "SetWindowBackgroundColorToDefaultColor";
-
-/// (macOS only).
-const _kSetWindowBackgroundColorToClear = "SetWindowBackgroundColorToClear";
-
-/// (macOS only).
-const _kSetBlurViewState = "SetBlurViewState";
-
-/// (macOS only).
-const _kAddVisualEffectSubview = "AddVisualEffectSubview";
-
-/// (macOS only).
-const _kUpdateVisualEffectSubviewProperties =
-    "UpdateVisualEffectSubviewProperties";
-
-/// (macOS only).
-const _kRemoveVisualEffectSubview = "RemoveVisualEffectSubview";
-
-/// (macOS only).
-const _kAddToolbar = "AddToolbar";
-
-/// (macOS only).
-const _kRemoveToolbar = "RemoveToolbar";
-
-/// (macOS only).
-const _kSetToolbarStyle = "SetToolbarStyle";
-
-/// (macOS only)
-const _kEnableShadow = "EnableShadow";
-
-/// (macOS only)
-const _kDisableShadow = "DisableShadow";
-
-/// (macOS only)
-const _kInvalidateShadows = "InvalidateShadows";
-
-/// (macOS only)
-const _kAddEmptyMaskImage = "AddEmptyMaskImage";
-
-/// (macOS only)
-const _kRemoveMaskImage = "RemoveMaskImage";
-
-/// (macOS only)
-const _kIgnoreMouseEvents = "IgnoreMouseEvents";
-
-/// (macOS only)
-const _kAcknowledgeMouseEvents = "AcknowledgeMouseEvents";
-
-/// (macOS only)
-const _kSetSubtitle = "SetSubtitle";
-
 final MethodChannel _kChannel = const MethodChannel(_kChannelName);
 final Completer<void> _kCompleter = new Completer<void>();
 
@@ -194,16 +52,26 @@ class Window {
   /// }
   /// ```
   static Future<void> initialize() async {
+    if (Platform.isMacOS) {
+      WindowManipulator.initialize();
+      setEffect(effect: WindowEffect.values[0]);
+      return;
+    }
+
     await _kChannel.invokeMethod(_kInitialize);
     _kCompleter.complete();
   }
 
   /// Sets specified effect for the window.
   ///
-  /// When using [WindowEffect.mica], [dark] argument can be used to switch between light or dark mode of Mica.
+  /// When using [WindowEffect.mica], [dark] argument can be used to switch
+  /// between light or dark mode of Mica.
   ///
-  /// When using [WindowEffect.acrylic], [WindowEffect.aero], [WindowEffect.disabled], [WindowEffect.solid] or [WindowEffect.transparent],
-  /// [color] argument can be used to change the resulting tint (or color) of the window background.
+  /// When using [WindowEffect.acrylic], [WindowEffect.aero],
+  /// [WindowEffect.disabled], [WindowEffect.solid] or
+  /// [WindowEffect.transparent],
+  /// [color] argument can be used to change the resulting tint (or color) of
+  /// the window background.
   ///
   /// _Examples_
   ///
@@ -226,6 +94,13 @@ class Window {
     Color color: Colors.transparent,
     bool dark: true,
   }) async {
+    if (Platform.isMacOS) {
+      final material =
+          WindowEffectToMaterialConverter.convertWindowEffectToMaterial(effect);
+      WindowManipulator.setMaterial(material);
+      return;
+    }
+
     await _kCompleter.future;
     await _kChannel.invokeMethod(
       _kSetEffect,
@@ -244,107 +119,122 @@ class Window {
 
   /// Hides window controls.
   static Future<void> hideWindowControls() async {
+    if (Platform.isMacOS) {
+      WindowManipulator.hideCloseButton();
+      WindowManipulator.hideMiniaturizeButton();
+      WindowManipulator.hideZoomButton();
+      return;
+    }
+
     await _kChannel.invokeMethod(_kHideWindowControls);
   }
 
   /// Shows window controls.
   static Future<void> showWindowControls() async {
+    if (Platform.isMacOS) {
+      WindowManipulator.showCloseButton();
+      WindowManipulator.showMiniaturizeButton();
+      WindowManipulator.showZoomButton();
+      return;
+    }
+
     await _kChannel.invokeMethod(_kShowWindowControls);
   }
 
   /// Makes the Flutter window fullscreen.
   static Future<void> enterFullscreen() async {
+    if (Platform.isMacOS) {
+      WindowManipulator.enterFullscreen();
+      return;
+    }
+
     await _kChannel.invokeMethod(_kEnterFullscreen);
   }
 
   /// Restores the Flutter window back to normal from fullscreen mode.
   static Future<void> exitFullscreen() async {
+    if (Platform.isMacOS) {
+      WindowManipulator.exitFullscreen();
+      return;
+    }
+
     await _kChannel.invokeMethod(_kExitFullscreen);
   }
 
   /// Gets the height of the titlebar.
   ///
   /// This value is used to determine the [[TitlebarSafeArea]] widget.
-  /// If the full-size content view is enabled, this value will be the height of the titlebar.
+  /// If the full-size content view is enabled, this value will be the height
+  /// of the titlebar.
   /// If the full-size content view is disabled, this value will be 0.
   /// This value is only available on macOS.
   static Future<double> getTitlebarHeight() async {
-    if (!Platform.isMacOS) {
-      throw new UnsupportedError(
-          'getTitlebarHeight() is only available on macOS.');
+    if (Platform.isMacOS) {
+      return WindowManipulator.getTitlebarHeight();
     }
 
-    await _kCompleter.future;
-    return await _kChannel.invokeMethod(_kGetTitlebarHeight);
+    throw new UnsupportedError(
+        'getTitlebarHeight() is only available on macOS.');
   }
 
   /// Sets the document to be edited.
   ///
-  /// This will change the appearance of the close button on the titlebar.
+  /// This will change the appearance of the close button on the titlebar:
+  ///
+  /// <img width="78" alt="image" src="https://user-images.githubusercontent.com/86920182/209436903-0a6c1f5a-4ab6-454f-a37d-78a5d699f3df.png">
+  ///
   /// This method is only available on macOS.
   static Future<void> setDocumentEdited() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetDocumentEdited);
+    WindowManipulator.setDocumentEdited();
   }
 
   /// Sets the document to be unedited.
   ///
   /// This method is only available on macOS.
   static Future<void> setDocumentUnedited() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetDocumentUnedited);
+    WindowManipulator.setDocumentUnedited();
   }
 
   /// Sets the represented file of the window.
   ///
   /// This method is only available on macOS.
   static Future<void> setRepresentedFilename(String filename) async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetRepresentedFile, {
-      'filename': filename,
-    });
+    WindowManipulator.setRepresentedFilename(filename);
   }
 
   /// Sets the represented URL of the window.
   ///
   /// This method is only available on macOS.
   static Future<void> setRepresentedUrl(String url) async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetRepresentedURL, {
-      'url': url,
-    });
+    WindowManipulator.setRepresentedUrl(url);
   }
 
   /// Hides the titlebar of the window.
   ///
   /// This method is only available on macOS.
   static Future<void> hideTitle() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kHideTitle);
+    WindowManipulator.hideTitle();
   }
 
   /// Shows the titlebar of the window.
   ///
   /// This method is only available on macOS.
   static Future<void> showTitle() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kShowTitle);
+    WindowManipulator.showTitle();
   }
 
   /// Makes the window's titlebar transparent.
   ///
   /// This method is only available on macOS.
   static Future<void> makeTitlebarTransparent() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kMakeTitlebarTransparent);
+    WindowManipulator.makeTitlebarTransparent();
   }
 
   /// Makes the window's titlebar opaque.
   ///
   /// This method is only available on macOS.
   static Future<void> makeTitlebarOpaque() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kMakeTitlebarOpaque);
+    WindowManipulator.makeTitlebarOpaque();
   }
 
   /// Enables the window's full-size content view.
@@ -354,66 +244,67 @@ class Window {
   /// the titlebar transparent.
   /// This method is only available on macOS.
   static Future<void> enableFullSizeContentView() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kEnableFullSizeContentView);
+    WindowManipulator.enableFullSizeContentView();
   }
 
   /// Disables the window's full-size content view.
   ///
   /// This method is only available on macOS.
   static Future<void> disableFullSizeContentView() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kDisableFullSizeContentView);
+    WindowManipulator.disableFullSizeContentView();
   }
 
   /// Zooms the window.
   ///
   /// This method is only available on macOS.
   static Future<void> zoomWindow() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kZoomWindow);
+    WindowManipulator.zoomWindow();
   }
 
   /// Unzooms the window.
   ///
   /// This method is only available on macOS.
   static Future<void> unzoomWindow() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kUnzoomWindow);
+    WindowManipulator.unzoomWindow();
   }
 
   /// Returns if the window is zoomed.
   ///
   /// This method is only available on macOS.
   static Future<bool> isWindowZoomed() async {
+    if (Platform.isMacOS) {
+      return WindowManipulator.isWindowZoomed();
+    }
+
     if (!Platform.isMacOS) {
       throw new UnsupportedError(
           'isWindowZoomed() is only available on macOS.');
     }
 
-    await _kCompleter.future;
-    return await _kChannel.invokeMethod(_kIsWindowZoomed);
+    return false;
   }
 
   /// Returns if the window is fullscreened.
   ///
   /// This method is only available on macOS.
   static Future<bool> isWindowFullscreened() async {
+    if (Platform.isMacOS) {
+      return WindowManipulator.isWindowFullscreened();
+    }
+
     if (!Platform.isMacOS) {
       throw new UnsupportedError(
           'isWindowFullscreened() is only available on macOS.');
     }
 
-    await _kCompleter.future;
-    return await _kChannel.invokeMethod(_kIsWindowFullscreened);
+    return false;
   }
 
   /// Hides the window's zoom button.
   ///
   /// This method is only available on macOS.
   static Future<void> hideZoomButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kHideZoomButton);
+    WindowManipulator.hideZoomButton();
   }
 
   /// Shows the window's zoom button.
@@ -421,16 +312,14 @@ class Window {
   /// The zoom button is visible by default.
   /// This method is only available on macOS.
   static Future<void> showZoomButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kShowZoomButton);
+    WindowManipulator.showZoomButton();
   }
 
   /// Hides the window's miniaturize button.
   ///
   /// This method is only available on macOS.
   static Future<void> hideMiniaturizeButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kHideMiniaturizeButton);
+    WindowManipulator.hideMiniaturizeButton();
   }
 
   /// Shows the window's miniaturize button.
@@ -438,16 +327,14 @@ class Window {
   /// The miniaturize button is visible by default.
   /// This method is only available on macOS.
   static Future<void> showMiniaturizeButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kShowMiniaturizeButton);
+    WindowManipulator.showMiniaturizeButton();
   }
 
   /// Hides the window's close button.
   ///
   /// This method is only available on macOS.
   static Future<void> hideCloseButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kHideCloseButton);
+    WindowManipulator.hideCloseButton();
   }
 
   /// Shows the window's close button.
@@ -455,8 +342,7 @@ class Window {
   /// The close button is visible by default.
   /// This method is only available on macOS.
   static Future<void> showCloseButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kShowCloseButton);
+    WindowManipulator.showCloseButton();
   }
 
   /// Enables the window's zoom button.
@@ -464,16 +350,14 @@ class Window {
   /// The zoom button is enabled by default.
   /// This method is only available on macOS.
   static Future<void> enableZoomButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kEnableZoomButton);
+    WindowManipulator.enableZoomButton();
   }
 
   /// Disables the window's zoom button.
   ///
   /// This method is only available on macOS.
   static Future<void> disableZoomButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kDisableZoomButton);
+    WindowManipulator.disableZoomButton();
   }
 
   /// Enables the window's miniaturize button.
@@ -481,16 +365,14 @@ class Window {
   /// The miniaturize button is enabled by default.
   /// This method is only available on macOS.
   static Future<void> enableMiniaturizeButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kEnableMiniaturizeButton);
+    WindowManipulator.enableMiniaturizeButton();
   }
 
   /// Disables the window's miniaturize button.
   ///
   /// This method is only available on macOS.
   static Future<void> disableMiniaturizeButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kDisableMiniaturizeButton);
+    WindowManipulator.disableMiniaturizeButton();
   }
 
   /// Enables the window's close button.
@@ -498,51 +380,41 @@ class Window {
   /// The close button is enabled by default.
   /// This method is only available on macOS.
   static Future<void> enableCloseButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kEnableCloseButton);
+    WindowManipulator.enableCloseButton();
   }
 
   /// Disables the window's close button.
   ///
   /// This method is only available on macOS.
   static Future<void> disableCloseButton() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kDisableCloseButton);
+    WindowManipulator.disableCloseButton();
   }
 
   /// Gets whether the window is currently being resized by the user.
   ///
   /// This method is only available on macOS.
   static Future<bool> isWindowInLiveResize() async {
-    if (!Platform.isMacOS) {
-      throw UnsupportedError(
-          'isWindowInLiveResize() is only available on macOS.');
+    if (Platform.isMacOS) {
+      return WindowManipulator.isWindowInLiveResize();
     }
 
-    await _kCompleter.future;
-    return await _kChannel.invokeMethod(_kIsWindowInLiveResize);
+    throw UnsupportedError(
+        'isWindowInLiveResize() is only available on macOS.');
   }
 
   /// Sets the window's alpha value.
   ///
   /// This method is only available on macOS.
   static Future<void> setWindowAlphaValue(double value) async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetWindowAlphaValue, <String, dynamic>{
-      'value': value,
-    });
+    WindowManipulator.setWindowAlphaValue(value);
+    return;
   }
 
   /// Gets if the window is visible.
   ///
   /// This method is only available on macOS.
   static Future<bool> isWindowVisible() async {
-    if (!Platform.isMacOS) {
-      throw UnsupportedError('isWindowVisible() is only available on macOS.');
-    }
-
-    await _kCompleter.future;
-    return await _kChannel.invokeMethod(_kIsWindowVisible);
+    return WindowManipulator.isWindowVisible();
   }
 
   /// Sets the window background color to the default (opaque) window color.
@@ -550,8 +422,8 @@ class Window {
   /// This method mainly affects the window's titlebar.
   /// This method is only available on macOS.
   static Future<void> setWindowBackgroundColorToDefaultColor() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetWindowBackgroundColorToDefaultColor);
+    WindowManipulator.setWindowBackgroundColorToDefaultColor();
+    return;
   }
 
   /// Sets the window background color to clear.
@@ -559,28 +431,31 @@ class Window {
   /// This method mainly affects the window's titlebar.
   /// This method is only available on macOS.
   static Future<void> setWindowBackgroundColorToClear() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetWindowBackgroundColorToClear);
+    WindowManipulator.setWindowBackgroundColorToClear();
   }
 
   /// Sets the blur view state.
   ///
   /// This method is only available on macOS.
   static Future<void> setBlurViewState(MacOSBlurViewState state) async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetBlurViewState, <String, dynamic>{
-      'state': state.toString().split('.').last,
-    });
+    final visualEffectViewState = BlurViewStateToVisualEffectViewStateConverter
+        .convertBlurViewStateToVisualEffectViewState(state);
+    WindowManipulator.setNSVisualEffectViewState(visualEffectViewState);
   }
 
-  /// Adds a visual effect subview to the application's window and returns its ID.
+  /// Adds a visual effect subview to the application's window and returns its
+  /// ID.
   ///
   /// This method is only available on macOS.
   static Future<int> addVisualEffectSubview(
       VisualEffectSubviewProperties properties) async {
-    await _kCompleter.future;
-    return await _kChannel.invokeMethod(
-        _kAddVisualEffectSubview, properties.toMap());
+    if (Platform.isMacOS) {
+      final newProperties =
+          properties.toMacOSWindowUtilsVisualEffectSubviewProperties();
+      return WindowManipulator.addVisualEffectSubview(newProperties);
+    }
+
+    return -1;
   }
 
   /// Updates the properties of a visual effect subview.
@@ -589,12 +464,10 @@ class Window {
   static Future<void> updateVisualEffectSubviewProperties(
       int visualEffectSubviewId,
       VisualEffectSubviewProperties properties) async {
-    await _kCompleter.future;
-    await _kChannel
-        .invokeMethod(_kUpdateVisualEffectSubviewProperties, <String, dynamic>{
-      'visualEffectSubviewId': visualEffectSubviewId,
-      ...properties.toMap(),
-    });
+    final newProperties =
+        properties.toMacOSWindowUtilsVisualEffectSubviewProperties();
+    WindowManipulator.updateVisualEffectSubviewProperties(
+        visualEffectSubviewId, newProperties);
   }
 
   /// Removes a visual effect subview from the application's window.
@@ -602,35 +475,24 @@ class Window {
   /// This method is only available on macOS.
   static Future<void> removeVisualEffectSubview(
       int visualEffectSubviewId) async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kRemoveVisualEffectSubview, <String, dynamic>{
-      'visualEffectSubviewId': visualEffectSubviewId,
-    });
+    WindowManipulator.removeVisualEffectSubview(visualEffectSubviewId);
   }
 
   /// Overrides the brightness setting of the window (macOS only).
   static Future<void> overrideMacOSBrightness({
     required bool dark,
   }) async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(
-      _kOverrideMacOSBrightness,
-      {
-        'dark': dark,
-      },
-    );
+    WindowManipulator.overrideMacOSBrightness(dark: dark);
   }
 
   /// Adds a toolbar to the window (macOS only).
   static Future<void> addToolbar() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kAddToolbar, {});
+    WindowManipulator.addToolbar();
   }
 
   /// Removes the window's toolbar (macOS only).
   static Future<void> removeToolbar() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kRemoveToolbar, {});
+    WindowManipulator.removeToolbar();
   }
 
   /// Sets the window's toolbar style (macOS only).
@@ -645,22 +507,19 @@ class Window {
   /// ```
   static Future<void> setToolbarStyle(
       {required MacOSToolbarStyle toolbarStyle}) async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetToolbarStyle, {
-      'toolbarStyle': toolbarStyle.name,
-    });
+    final newToolbarStyle = MacOSToolbarStyleToWindowToolbarStyleConverter
+        .convertMacOSToolbarStyleToWindowToolbarStyle(toolbarStyle);
+    WindowManipulator.setToolbarStyle(toolbarStyle: newToolbarStyle);
   }
 
   /// Enables the window's shadow (macOS only).
   static Future<void> enableShadow() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kEnableShadow, {});
+    WindowManipulator.enableShadow();
   }
 
   /// Disables the window's shadow (macOS only).
   static Future<void> disableShadow() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kDisableShadow, {});
+    WindowManipulator.disableShadow();
   }
 
   /// Invalidates the window's shadow (macOS only).
@@ -668,8 +527,7 @@ class Window {
   /// This is a fairly technical method and is included here for
   /// completeness' sake. Normally, it should not be necessary to use it.
   static Future<void> invalidateShadows() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kInvalidateShadows, {});
+    WindowManipulator.invalidateShadows();
   }
 
   /// Adds an empty mask image to the window's view (macOS only).
@@ -681,14 +539,12 @@ class Window {
   /// enabled when using an empty mask image can cause visual artifacts
   /// and performance issues.
   static Future<void> addEmptyMaskImage() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kAddEmptyMaskImage, {});
+    WindowManipulator.addEmptyMaskImage();
   }
 
   /// Removes the window's mask image (macOS only).
   static Future<void> removeMaskImage() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kRemoveMaskImage, {});
+    WindowManipulator.removeMaskImage();
   }
 
   /// Makes a window fully transparent (with no blur effect) (macOS only).
@@ -717,8 +573,7 @@ class Window {
   /// may be desirable when used in conjunction with
   /// `Window.makeWindowFullyTransparent()`.
   static Future<void> ignoreMouseEvents() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kIgnoreMouseEvents, {});
+    WindowManipulator.ignoreMouseEvents();
   }
 
   /// Makes the window acknowledge mouse events (macOS only).
@@ -727,17 +582,13 @@ class Window {
   /// may be desirable when used in conjunction with
   /// `Window.makeWindowFullyTransparent()`.
   static Future<void> acknowledgeMouseEvents() async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kAcknowledgeMouseEvents, {});
+    WindowManipulator.acknowledgeMouseEvents();
   }
 
   /// Sets the subtitle of the window (macOS only).
   ///
   /// To remove the subtitle, pass an empty string to this method.
   static Future<void> setSubtitle(String subtitle) async {
-    await _kCompleter.future;
-    await _kChannel.invokeMethod(_kSetSubtitle, {
-      'subtitle': subtitle,
-    });
+    WindowManipulator.setSubtitle(subtitle);
   }
 }