+++ lib/window_effect.dart
@@ -17,15 +17,19 @@ enum WindowEffect {
   /// Works only on Windows.
   aero,
 
-  /// Acrylic is a type of brush that creates a translucent texture. You can apply acrylic to app surfaces to add depth and help establish a visual hierarchy.
+  /// Acrylic is a type of brush that creates a translucent texture. You can
+  /// apply acrylic to app surfaces to add depth and help establish a visual
+  /// hierarchy.
   /// Works only on Windows 10 version 1803 or higher.
   acrylic,
 
-  /// Mica is an opaque, dynamic material that incorporates theme and desktop wallpaper to paint the background of long-lived windows.
+  /// Mica is an opaque, dynamic material that incorporates theme and desktop
+  /// wallpaper to paint the background of long-lived windows.
   /// Works only on Windows 11 or greater.
   mica,
 
-  /// Tabbed is a Mica like material that incorporates theme and desktop wallpaper, but having more transparency.
+  /// Tabbed is a Mica like material that incorporates theme and desktop
+  /// wallpaper, but having more transparency.
   /// Works only on later Windows 11 versions (builds higher than 22523).
   tabbed,